# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/cats": {
      "post": {
        "operationId": "CatsController_create",
        "summary": "猫データを作成",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "猫データが正常に作成されました"
          },
          "400": {
            "description": "無効なデータです"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "CatsController_findAll",
        "summary": "猫データを検索・一覧取得",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "ページ番号",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "1ページあたりの件数",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "検索キーワード",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "breedId",
            "required": false,
            "in": "query",
            "description": "品種ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "coatColorId",
            "required": false,
            "in": "query",
            "description": "毛色ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gender",
            "required": false,
            "in": "query",
            "description": "性別",
            "schema": {
              "enum": [
                "MALE",
                "FEMALE",
                "NEUTER",
                "SPAY",
                "1",
                "2",
                "3",
                "4"
              ],
              "type": "string"
            }
          },
          {
            "name": "isInHouse",
            "required": false,
            "in": "query",
            "description": "在舎中フィルター",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "ageMin",
            "required": false,
            "in": "query",
            "description": "最小年齢",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "ageMax",
            "required": false,
            "in": "query",
            "description": "最大年齢",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "ソート項目",
            "schema": {
              "default": "createdAt",
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "ソート順",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "ステータス",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "猫データの一覧"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/cats/statistics": {
      "get": {
        "operationId": "CatsController_getStatistics",
        "summary": "猫データの統計情報を取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "統計情報"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/cats/kittens": {
      "get": {
        "operationId": "CatsController_findKittens",
        "summary": "子猫一覧を取得（生後6ヶ月未満、母猫ごとにグループ化）",
        "parameters": [
          {
            "name": "motherId",
            "required": false,
            "in": "query",
            "description": "母猫ID（指定時はその母猫の子猫のみ取得）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "ページ番号",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "1ページあたりの件数",
            "schema": {
              "default": 50,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "検索キーワード（子猫名）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "ソート項目",
            "schema": {
              "default": "birthDate",
              "example": "birthDate",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "ソート順",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "子猫データの一覧（母猫ごとにグループ化）"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/cats/genders": {
      "get": {
        "operationId": "CatsController_getGenders",
        "summary": "性別マスタデータを取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "性別マスタデータを返却"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/cats/{id}": {
      "get": {
        "operationId": "CatsController_findOne",
        "summary": "IDで猫データを取得",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "猫データのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "猫データ"
          },
          "404": {
            "description": "猫データが見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "CatsController_update",
        "summary": "猫データを更新",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "猫データのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "猫データが正常に更新されました"
          },
          "400": {
            "description": "無効なデータです"
          },
          "404": {
            "description": "猫データが見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CatsController_remove",
        "summary": "猫データを削除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "猫データのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "猫データが正常に削除されました"
          },
          "404": {
            "description": "猫データが見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/cats/{id}/breeding-history": {
      "get": {
        "operationId": "CatsController_getBreedingHistory",
        "summary": "猫の繁殖履歴を取得",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "猫データのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "繁殖履歴"
          },
          "404": {
            "description": "猫データが見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/cats/{id}/care-history": {
      "get": {
        "operationId": "CatsController_getCareHistory",
        "summary": "猫のケア履歴を取得",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "猫データのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ケア履歴"
          },
          "404": {
            "description": "猫データが見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/cats/{id}/family": {
      "get": {
        "operationId": "CatsController_getCatFamily",
        "summary": "猫の家族情報を取得（血統タブ用）",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "猫データのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "家族情報（親・兄弟姉妹・子猫）"
          },
          "404": {
            "description": "猫データが見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/cats/{id}/weight-records": {
      "get": {
        "operationId": "CatsController_getWeightRecords",
        "summary": "猫の体重記録一覧を取得",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "猫データのID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "ページ番号",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "1ページあたりの件数",
            "schema": {
              "default": 50,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "開始日",
            "schema": {
              "example": "2024-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "終了日",
            "schema": {
              "example": "2024-12-31",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "ソート順",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "体重記録一覧"
          },
          "404": {
            "description": "猫データが見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "CatsController_createWeightRecord",
        "summary": "猫の体重記録を追加",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "猫データのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeightRecordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "体重記録が正常に作成されました"
          },
          "400": {
            "description": "無効なデータです"
          },
          "404": {
            "description": "猫データが見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/cats/weight-records/{recordId}": {
      "get": {
        "operationId": "CatsController_getWeightRecord",
        "summary": "体重記録を取得",
        "parameters": [
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "description": "体重記録のID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "体重記録"
          },
          "404": {
            "description": "体重記録が見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "CatsController_updateWeightRecord",
        "summary": "体重記録を更新",
        "parameters": [
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "description": "体重記録のID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWeightRecordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "体重記録が正常に更新されました"
          },
          "400": {
            "description": "無効なデータです"
          },
          "404": {
            "description": "体重記録が見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CatsController_deleteWeightRecord",
        "summary": "体重記録を削除",
        "parameters": [
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "description": "体重記録のID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "体重記録が正常に削除されました"
          },
          "404": {
            "description": "体重記録が見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/cats/weight-records/bulk": {
      "post": {
        "operationId": "CatsController_createBulkWeightRecords",
        "summary": "複数の猫の体重を一括登録",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBulkWeightRecordsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "体重記録が正常に一括作成されました"
          },
          "400": {
            "description": "無効なデータです"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Cat Management System API",
    "description": "API for managing cat breeding and care records",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCatDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "猫の名前",
            "example": "Alpha"
          },
          "gender": {
            "type": "string",
            "description": "性別",
            "enum": [
              "MALE",
              "FEMALE",
              "NEUTER",
              "SPAY"
            ],
            "example": "MALE"
          },
          "birthDate": {
            "type": "string",
            "description": "生年月日",
            "example": "2024-05-01"
          },
          "breedId": {
            "type": "string",
            "description": "品種ID"
          },
          "coatColorId": {
            "type": "string",
            "description": "毛色ID"
          },
          "microchipNumber": {
            "type": "string",
            "description": "マイクロチップ番号"
          },
          "registrationNumber": {
            "type": "string",
            "description": "登録番号"
          },
          "description": {
            "type": "string",
            "description": "説明・備考"
          },
          "isInHouse": {
            "type": "boolean",
            "description": "施設内に在舎しているか"
          },
          "fatherId": {
            "type": "string",
            "description": "父猫のID"
          },
          "motherId": {
            "type": "string",
            "description": "母猫のID"
          },
          "tagIds": {
            "description": "タグID配列",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "gender",
          "birthDate"
        ]
      },
      "UpdateCatDto": {
        "type": "object",
        "properties": {}
      },
      "CreateWeightRecordDto": {
        "type": "object",
        "properties": {
          "weight": {
            "type": "number",
            "description": "体重（グラム単位）",
            "example": 350,
            "minimum": 1,
            "maximum": 50000
          },
          "recordedAt": {
            "type": "string",
            "description": "記録日時（省略時は現在日時）",
            "example": "2024-01-15T10:00:00.000Z"
          },
          "notes": {
            "type": "string",
            "description": "メモ",
            "example": "ミルクをよく飲んでいる"
          }
        },
        "required": [
          "weight"
        ]
      },
      "UpdateWeightRecordDto": {
        "type": "object",
        "properties": {
          "weight": {
            "type": "number",
            "description": "体重（グラム単位）",
            "example": 380,
            "minimum": 1,
            "maximum": 50000
          },
          "recordedAt": {
            "type": "string",
            "description": "記録日時",
            "example": "2024-01-15T10:00:00.000Z"
          },
          "notes": {
            "type": "string",
            "description": "メモ",
            "example": "順調に成長中"
          }
        }
      },
      "BulkWeightRecordItemDto": {
        "type": "object",
        "properties": {
          "catId": {
            "type": "string",
            "description": "猫ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "weight": {
            "type": "number",
            "description": "体重（グラム単位）",
            "example": 350,
            "minimum": 1,
            "maximum": 50000
          },
          "notes": {
            "type": "string",
            "description": "メモ",
            "example": "元気"
          }
        },
        "required": [
          "catId",
          "weight"
        ]
      },
      "CreateBulkWeightRecordsDto": {
        "type": "object",
        "properties": {
          "recordedAt": {
            "type": "string",
            "description": "記録日時（全レコード共通）",
            "example": "2024-01-15T10:00:00.000Z"
          },
          "records": {
            "description": "体重記録の配列",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BulkWeightRecordItemDto"
            }
          }
        },
        "required": [
          "recordedAt",
          "records"
        ]
      },
...
    }
  }
}